requirements-dev.txt
aiohttp-devtools
flake8
pylint
pytest-aiohttp
pytest
